Update connection stuct to match requirement
In a normal workflow the connection state should be predictable and
there should be an easy way to get it.

This commit adds a goroutine to watch the connection state and trigger
listeners accordingly. Any change to the connection state is passed to
the state channel and, in turn, will create as many goroutines as there
is listener.