Align implementation with other libraries
Our java library overloads Get and History to provide a way to load the
next page from a paginated query (using the URL only). We could provide
a second method to avoid overloading (such as GetNextPage) which would
take a URL as a parameter like java does, but that would be a Go library
only method...

On the other hand, the ruby library returns an object that is enumerable
and which can query the Rest API directly...

All in all this is the closest approach we can go with. Return a
PaginatedPresenceMessages and let it query new results. Making it
effectively an immutable that will never change.