bug: unpinned range scope variables in tests
Ran https://github.com/kyoh86/scopelint to find unpinned variables and corrected
them.

When a closure is used inside of a goroutine the closure is bound once and the
variable can get out of sync. Variables declared inside the body of a loop are
not shared between iterations, so we pin the value inside the loop body to
prevent this bug from happening. See
https://github.com/golang/go/wiki/CommonMistakes#using-goroutines-on-loop-iterator-variables
for more details.