bug: TestRestClient fails, race cond with sandbox
The Stats test sends a POST request to the ably sandbox, and then follows up
with a GET request. If the GET request happens too quickly after the POST, the
sandbox will not have populated data yet and the Stats request will return an
empty response which causes the test to fail.

A loop is added that will call client.Stats every 500ms and check for a
nonempty result. If after 10s no result is present the test will
timeout.