Remove leaky channel's connection event listener goroutine.
This goroutine was never killed, and there's really no proper place to
kill it. It should live as long as the channel instance lives, but the
channel instance is garbage-collected.

We could use runtime.SetFinalizer to tie the goroutine's lifetime to
the instance's lifetime, but it's simpler to just not have the goroutine
in the first place.

So we replace the channel-based broadcast with callback-based broadcast,
so that the listener (now just a function) is garbage-collected too.

Eventually, we can use the callback-based EventEmitter to cover this
in a cleaner way. This should be part of https://github.com/ably/ably-go/pull/144.