RTL2: Remove old State, replacing with EventEmitter everywhere.
Unfortunately this isn't easy to do in more than one step, so it's a big
refactor, including tests.

The eventEmitter public, type-safe wrappers for Connection and
RealtimeChannel have been moved to a separate type to be embedded. This
allows us to reuse it for the internal event emitter too.

Now the locks are kept inside RealtimeConn and RealtimeChannel.

Old State.set has been renamed to lockSet, and syncSet to set. This is
consistent with other places where the "lock" prefix indicates that it
should be called with the mutex already locked.

Test recorders have been substituted by simpler ConnStates and
ChanStates.