Use ablytest.IsTrue for tests that may not be true right away.
IsTrue is more like the old behavior of those tests. Since EventEmitter
is concurrent with synchronous test code, there's a race between
receiving the expected events and the point where we check the events
have been received. So wait for a bit before giving up.

Also, remove part of TestRealtime_RealtimeHost that was racy but wasn't
really testing anything useful.