Propagate contexts everywhere.
Now HTTP requests require contexts. Since lots of things end up making
HTTP requests, contexts need to be propagated everywhere. Huge change,
but not one that can be easily or usefully split.

For where we don't have a context passed as a parameter because it's the
library itself who initiates the operation, a base context can be set
with a new ClientOption.