TestFixConnLeak_ISSUE89: Change the way it detects closed conns
In the new CI, we often see this:

--- FAIL: TestFixConnLeak_ISSUE89 (6.87s)
    rest_client_test.go:574: expected conn 10.1.0.4:47296 to be closed
    rest_client_test.go:574: expected conn 10.1.0.4:47308 to be closed
    rest_client_test.go:574: expected conn 10.1.0.4:47312 to be closed

It's probably just a race condition, so, fix that by waiting a bit.

Also, the way it was checking that conns were closed was by trying to
reuse them, seeing if that failed with a particular error message.
That's too implementation-dependent. Now it checks that the Close()
method is called on the connections.