Remove flaky TestRealtimeChannel_Detach.
It first subscribed to a channel, then published, then detaches from the
channel, then checks that the published messsage is received.

But there's no guarantee of this. Ably will send us the published
message asynchronously at some point, concurrently with us detaching,
after which Ably won't send us the published message.

Whatever else the test is trying to check, it is or will be covered by
spec-oriented tests anyway.

Fixes #309.