Unify PresenceAction types.
We had one for presence events and another for presence messages.

Now we use the one for presence messages for both. It is defined
differently to other EventEmitter types, so it's now inconsistent in
that sense, but the alternative would be converting between both and
that would be too complex.