examples: Remove go.mod and go.sum
Given that these examples are in the same git repository as the ably-go
package, I think it's reasonable to expect them to run against the
version of ably-go as defined in the currently checked out commit,
rather than expecting them to run against a released version.

If someone is interested in examples that work with a specific release,
then they can view the examples as they were in that git tag.

For this reason, there is no reason to have a go.mod in the examples
directory, so this commit removes it.

Signed-off-by: Lewis Marshall <lewis.marshall@ably.com>