Add connection key to messages.
This fixes #422 a customer previously raised a PR to add ConnectionKey to
the message. However the value of this new connection key field was never
populated, so it still was not possible for a rest client to publish on
behalf of a realtime client. This PR populates the connection key when rest
message are published.