Handle fallback in Request()
My interpretation of RSC19e is that Request() should only return an
error if there's a network error. However it also states the standard
fallback rulles apply.

So we save the response then check it's status. If the status we return
an error so that the fallback functions trigger. If we end up with an
error still after exausting all fallbacks then we return the last req
that succeeded.

This lets us try fallbacks and also return bad requests for the user to
inspect.