Support msgpack in Reqeuest()
This used to work before 5574cbd where request was hard coded to return
a map.

Now the user specifies the type via the out argument to Item(). So the
decoding was delayed by storing the value in a json.RawMessage, then
done for real during the Item() call.

This expands that by storing the data in a Raw type that defers to
json.RawMessage or codec.Raw depending on the encoding.