Fix message extras unmarshaled to incompatible json type
Message extras should be a json compatible map[string]interface{}, but
when marshalling to msgpack and back again, any nested objects become
map[interface{}]interface{} which is incompatible with json marshalling.

Set the map type on the msgpack handle to map[string]interface{} so that
nested objects are correctly unmarshaled.