feat: Close old connection when new ones set
Prevoiusly old connections were just overwritten as a new connections
had been established. This was causing a go-routine leak as the old
connections still had running eventloops.

This commit now asyncrously closes these connections so the `Connection`
type is not blocked waiting for an old connection to close.

The websocket library has a 5 second timeout on closing old connections
so no timeout logic has been a applied to the new go routine closing
the old connection.