fix: stop setting readlimit on a connected event
This commit removes setting the websocket readlimit based on
the `ConnectionDetails.MaxMessageSize` received when a message
with a Connected action (4).

The `ConnectionDetails.MaxMessageSize` should be used to restrict
publishing messages over this value.

The Connection.isReadLimitSetExternally has been removed as this
is no longer required.

Connections do have a read limit set still but the default limit
has been raised to 2mb.